requirements.txt
requests
pandas
pygwalker
python-dotenv 